\name{CovariatesAnalysis}
\alias{CovariatesAnalysis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Analyze covariates
}
\description{
Analyze covariates
}
\usage{
CovariatesAnalysis(disease,studyFolder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{disease}{Name of disease want to analyze}
  \item{studyFolder}{Name of local folder to place results; make sure to use forward slashes(/)}
}
\details{
This function creates files specifying the analyses that will be performed.}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
