\name{ExtractCovariates}
\alias{ExtractCovariates}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extract covariates and save cohorts
}
\description{
Extract covariates and save cohorts
}
\usage{
ExtractCovariates(connectionDetails, cdmDatabaseSchema, studyFolder)
}

\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{studyFolder}{Name of local folder to place results; make sure to use forward slashes(/)}

\item{cdmVersion}{cdmVersion 4/5}

\item{disease}{Name of disease want to analyze}

\item{cohortId}{cohortId of interest disease generated by Atlas}
}
\details{
This function creates files specifying the analyses that will be performed.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
